data-ingestion/requirements.txt
kafka-python==2.0.2
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
redis==5.0.1
pydantic==2.5.2
loguru==0.7.2
aiohttp==3.9.1
websockets==12.0